detect/http_raw_header: move tests into tests/
