+++ src/detect-engine-file.c
@@ -33,7 +33,6 @@
 #include "detect-filestore.h"
 
 #include "detect-engine-hcbd.h"
-#include "detect-engine-hrhd.h"
 #include "detect-engine-dcepayload.h"
 #include "detect-engine-file.h"
